// frontend/src/client/types.gen.ts
export type DealPublic = {
    name: string;
    amount?: (number | null);
    stage?: DealStage;
    probability?: (number | null);
    expected_close_date?: (string | null);
    notes?: (string | null);
    id: string;
    company_id: string;
    contact_id: (string | null);
    owner_id: string;
    tenant_id: string;
    created_at?: (string | null);
    updated_at?: (string | null);
};

export type DealsPublic = {
    data: Array<DealPublic>;
    count: number;
};

export type DealStage = 'lead' | 'qualified' | 'proposal' | 'negotiation' | 'closed_won' | 'closed_lost';